+++ GhostEXIF/MetadataManager.swift
@@ -466,7 +466,9 @@ class MetadataManager: ObservableObject {
                 bitsPerComponent: 8,
                 bytesPerRow: 0,
                 space: colorSpace,
-                bitmapInfo: CGImageAlphaInfo.premultipliedLast.rawValue
+                bitmapInfo: (opaque
+                    ? CGImageAlphaInfo.noneSkipLast
+                    : CGImageAlphaInfo.premultipliedLast).rawValue
               ) else { return nil }
 
         if opaque {